feat: :sparkles: reload config when changes detected
Uses the file watcher to watch the config directory for changes to our config files. When it detects a file is saved or changes sizes, it takes a hash of the file, compares it to the previous hash, and if they are different it reloads the config. This basically allows for hot config reloading as soon as you update one of the config files!

#61