feat: :zap: save records database in config directory
Save the leaderboard/records database in the configuration directory instead of the root folder for the game. In some cases, the records.db was getting erased or lost when the folder was cleaned or reset (e.g. when running in a docker container). The configuration directory should be safe from that because if it isn't you would be reconfiguring every update.

Moves where the records.db is saved at. #106