refactor: 🔥 clean up logic determining send to discord
Moved the code which used to be under the individual patches into a single Handlers class which takes care of checking configs for settings, recording stats in the database, and sending to discord. This makes the patch code much easier to read. Also cleaned up the logic in the Handlers with guards instead of duplicated code in if/else blocks.