feat: ⚡ change to using local assemblies
Using publicized assemblies locally is the preferred method for modding Valheim, to ensure using the correct versions for the target release of the game. It's also "best practice" to use the BepInEx dependencies from the Valheim BepInEx package instead of using the upstream BepInEx package.  The Valheim BepInEx package includes unstripped Unity assemblies as well.